Daisy -AI Assistant- 0.5/Daisy -AI Assistant- 0.2 copy/agent-controller/requirements.txt
watchdog>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
gTTS>=2.4.0
playsound>=1.3.0
SpeechRecognition>=3.10.0
openai>=1.0.0
pyaudio>=0.2.14
pydub>=0.25.1
groq>=0.34.0
requests>=2.31.0

